+++ packages/angular-library/projects/components/src/public-api.ts
@@ -1,6 +1,7 @@
 /*
- * Public API Surface of component-library
+ * Public API Surface of components
  */
 
-export * from './lib/components.service';
-export * from './lib/components.component';
+export * from './lib/components.module';
+export { DIRECTIVES } from './lib/stencil-generated';
+export * from './lib/stencil-generated/components';