+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export { Components, JSX } from './components';
+export type * from './components';
 
 // all other exports (used in documentation)
 export * from './exports';