+++ packages/vue-library/lib/plugin.ts
@@ -1,0 +1,8 @@
+import { Plugin } from 'vue';
+import { defineCustomElements } from '@dytesdk/ui-kit/loader';
+
+export const ComponentLibrary: Plugin = {
+  async install() {
+    defineCustomElements();
+  },
+};